mega update
now it's possible to have up to 4 players in the same game (obviously not online but on the same screen), not the best key bindings for now (probably i'll make the player colors and key bindings customizable). debug feature at this point is very performance-heavy. It's possible to customize the number of the players in the URL string by changing the number after the "&players=" option